updating authentication to work with laravel 5.1
